requirements.txt
numpy==1.21.2
tensorboard==2.6.0
torchvision==0.10.0
